django-be/requirements.txt
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==24.1.0
Babel==2.15.0
bcc==0.18.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.1.0
blinker==1.4
Brlapi==0.8.3
build==1.2.2.post1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
comm==0.2.2
command-not-found==0.3
cookiecutter==2.6.0
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
debugpy==1.8.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro==1.7.0
Django==5.1.2
django-cors-headers==4.5.0
django-ninja==1.3.0
djangorestframework==3.15.2
dynaconf==3.2.6
exceptiongroup==1.2.2
executing==2.0.1
fastjsonschema==2.20.0
fqdn==1.5.1
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
hidpidaemon==18.4.6
httpcore==1.0.5
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.26.0
ipywidgets==8.1.3
isoduration==20.11.0
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.4
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.11
kedro==0.19.9
kedro-telemetry==0.6.1
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.20.0
macaroonbakery==1.3.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
more-itertools==10.5.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
notebook==7.2.1
notebook_shim==0.2.4
numpy==2.1.2
oauthlib==3.2.0
omegaconf==2.3.0
open-fprintd==0.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.3
pandocfilters==1.5.1
parse==1.20.2
parso==0.8.4
pexpect==4.9.0
platformdirs==4.2.2
pluggy==1.5.0
pop-transition==1.1.2
pre_commit_hooks==5.0.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
protobuf==3.12.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pycairo==1.20.1
pycparser==2.22
pycryptodome==3.21.0
pycups==2.0.1
pydantic==2.9.2
pydantic_core==2.23.4
pydbus==0.6.0
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproject_hooks==1.2.0
pyRFC3339==1.1
PySocks==1.7.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-json-logger==2.0.7
python-slugify==8.0.4
python-validity==0.13
python-xlib==0.29
pytoolconfig==1.3.1
pytz==2022.1
pyusb==1.2.1.post1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==26.1.0
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.35.1
repolib==2.2.1
repoman==1.4.0
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.3
rope==1.13.0
rpds-py==0.19.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
scrapingbee==2.0.1
SecretStorage==3.3.1
selenium==4.24.0
Send2Trash==1.8.3
sessioninstaller==0.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.5.1
stack-data==0.6.3
systemd-python==234
terminado==0.18.1
text-unidecode==1.3
tinycss2==1.3.0
toml==0.10.2
tomli==2.0.1
tornado==6.4.1
traitlets==5.14.3
trio==0.26.2
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-drivers-common==0.0.0
ufw==0.36.1
uri-template==1.3.0
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
widgetsnbextension==4.0.11
wsproto==1.2.0
xdg==5
xkit==0.0.0
zipp==1.0.0
